Patch required to build SDL against the latest ppapi  headers.
This works around the problem that more recent
ppapi version have moved the declaration of CallbackFactories 
form:
ppapi/cpp/completion_callback.h
to the new file:
ppapi/utility/completion_callback_factory.h

This causes quite a bit of headache in nacl-ports
as we build it with a large variety of SDKs which
all have different sets of ppapi headers.

For now we (ab)use a versioned symbol  from pp_var to control the
the include behavior.

This could be made somewhat cleaner if ppapi exported 
a number symbol based on the say the svn rev, e.g.
#define PPPAI_VERSION_CONTROL_NO 9123456

which could be use in nacl-ports lik so:
#if  PPPAI_VERSION_CONTROL_NO > x



BUG= http://code.google.com/p/naclports/issues/detail?id=27
Review URL: https://chromiumcodereview.appspot.com/9148047

git-svn-id: https://naclports.googlecode.com/svn/trunk/src@519 7dad1e8b-422e-d2af-fbf5-8013b78bd812