Rename naclports packages not to include version.
This change renames the directories containing packages
to that they do not contain version numbers.  It also
renames the .sh build script and the .patch files likewise.
The rational here is that:
 a) its redundant information
 b) it makes it hard to navigate the tree and run scripts.
    (./libraries/lua/nacl-lua.sh is much easier to type
     than ./libraries/lua-5.4.1/nacl-lua-5.4.1.sh)
 c) its makes it hard to roll package versions (you need to
    rename folders and then rename files within them
    to bump a version)

One downside is that it becomes messier when you want to
package two versions of the same library.  This would then
be and exception to the rule when a version could be
included in the package name (similar to how it is done
for deb packages).

BUG=

Review URL: https://codereview.chromium.org/11926029

git-svn-id: https://naclports.googlecode.com/svn/trunk/src@682 7dad1e8b-422e-d2af-fbf5-8013b78bd812