Fix nacl_build.mk build system.
This build system which is used by several projects in the
examples folder looks like it hasn't worked at all for a
long time.  This change makes the build system work but
many of the examples that use it don't compile against the
latest PPAPI headers.

BUG=None
TEST=run 'make' for each example

Review URL: https://codereview.chromium.org/12051092

git-svn-id: https://naclports.googlecode.com/svn/trunk/src@689 7dad1e8b-422e-d2af-fbf5-8013b78bd812