Use git for patch management in individual ports.
Each time we extract an upstream archive and apply
patches to it, do this in git.  This makes it much
easier to work with ports and track change locally.

Also, remove the need to Zip and Bz2 specific versions
of the shell code for downloading and extracting archives.
This can be instead inferred from the archive filename.

Also, allow package script to set CONFIG_SUB to point
to a non-standard config.sub location.

BUG=None
R=binji@chromium.org

Review URL: https://codereview.chromium.org/15007009

git-svn-id: https://naclports.googlecode.com/svn/trunk/src@765 7dad1e8b-422e-d2af-fbf5-8013b78bd812