Add initial SDL opengl support based on Regal.
This is non-functional at the momemt but is needed to fix
the build.  Without this the SDL opengl tests will detect
and use the mese headers which are also part of naclports,
but fail to link.

Also, add Regal to the same build shard as SDL.

R=binji@chromium.org

Review URL: https://codereview.chromium.org/16679008

git-svn-id: https://naclports.googlecode.com/svn/trunk/src@784 7dad1e8b-422e-d2af-fbf5-8013b78bd812