Add initial python port.
This include the python source code itself which is used
to build the python library and the sel_ldr version of
the python executable.    I've also added a python_ppapi
project which is just a ppapi_simple-based application
that run the main python entry point.

BUG=
R=binji@chromium.org

Review URL: https://codereview.chromium.org/17990002

git-svn-id: https://naclports.googlecode.com/svn/trunk/src@802 7dad1e8b-422e-d2af-fbf5-8013b78bd812