Enable use of goma when building naclports
This is achieved by conditionally adding 'gomacc' to
the start of NACLCC and NACLCXX.  This is only done
if NACL_GOMA is set to 1 and the goma test for the
compiler succeeds.

Mostly this change only required the quoting of the
CC and CXX variables which can now contain spaces but
it also involved changes to the way cmake builds pass
the compiler name to cmake.

BUG=
R=binji@chromium.org

Review URL: https://codereview.chromium.org/16553004

git-svn-id: https://naclports.googlecode.com/svn/trunk/src@803 7dad1e8b-422e-d2af-fbf5-8013b78bd812