Fixes for newlib python build.
Python (under newlib anyway) seems to rely on a working
getcwd() and the one from -lnosys is not enough.  Always
returning "." seems to be enough to get the interpreter
working.

Statically link all the common posix modules in
the newlib build.

Define "timezone" and "daylight" which are expected
by the time module.

Increase the buffer size used to report the compiler
used to build python since the nacl compiler string
is very long (it includes a git hash!).

When a posix file i/o function returns ENOSYS report
the function name that reported it rather then that
filename that the function was called with (we should
probably upstream this).

BUG= http://code.google.com/p/nativeclient/issues/detail?id=1229
R=binji@chromium.org

Review URL: https://codereview.chromium.org/19503002

git-svn-id: https://naclports.googlecode.com/svn/trunk/src@829 7dad1e8b-422e-d2af-fbf5-8013b78bd812