Don't remove entire out directory on 'make clean'
Only remove certain direcotires from the out direcotry
on 'make clean'.  In paritular, don't remove the cache
of downloaded tarballs, the publish folder, or (in the
case of the bots) the nacl sdk root itself.

'make reallyclean' can be used to remove the absolutely
everything if needed.

TEST=./buildbot_selector.sh
R=binji@chromium.org

Review URL: https://codereview.chromium.org/22808005

git-svn-id: https://naclports.googlecode.com/svn/trunk/src@857 7dad1e8b-422e-d2af-fbf5-8013b78bd812