Make gcc error out if -shared is used with newlib.
Unfortunately the old version of gcc that we ship for x85-newlib
produced 257Mb files when -shared is used.  In fact, -shared is
not suppored.  This change modifies the specs file of the compiler
to produce an error message if -shared is passed to the linker.

R=binji@chromium.org

Review URL: https://codereview.chromium.org/22914010

git-svn-id: https://naclports.googlecode.com/svn/trunk/src@866 7dad1e8b-422e-d2af-fbf5-8013b78bd812