Cleanup fontconfig patch file.
This reduces significantly the size of the fontconfig patch
file.

Rather than including all the auto-generated files in the
patch we now do a host build first in order to generate the
files, and then use the host-generated files in the
cross-build.

We should probably investigate updating the upstream version
to see if they fixed cross building at some point.

R=binji@chromium.org

Review URL: https://codereview.chromium.org/22861034

git-svn-id: https://naclports.googlecode.com/svn/trunk/src@887 7dad1e8b-422e-d2af-fbf5-8013b78bd812