Fix sqlite glibc build.
Add -pthread of sqlite link line since nacl_io now depends
on it. Really the nacl_io.so should declare its own
dependence on pthread but that wouldn't be enough to make
the .a file useable.

Also, fix the Ruby build under pnacl which was trying to
create was calling WriteSelLdrScript on a non-existant file.

TBR=binji

Review URL: https://codereview.chromium.org/23628013

git-svn-id: https://naclports.googlecode.com/svn/trunk/src@903 7dad1e8b-422e-d2af-fbf5-8013b78bd812