Fix openssl "make install".  Actually popd...
A previous CL: https://codereview.chromium.org/37463005/
broke the "make install" step. When adding the unit test
running, we has "pushd test" and a "popd test". This
left us in the "test" directory when doing the "make install"
which didn't work.  Change the "popd test" to "popd" so that
the "make install" runs in the correct directory.

Also, for newlib, create a separate Makefile that's a stub
so that "make install" won't fail to link tests that don't
link.

R=sbc@chromium.org

Review URL: https://codereview.chromium.org/53493002

git-svn-id: https://naclports.googlecode.com/svn/trunk/src@957 7dad1e8b-422e-d2af-fbf5-8013b78bd812