TeX Live 2013 build in examples
TeX Live currently only builds with glibc; we need to add some of the
missing function definition into toolchain provided headers before we
can build with newlib.

The configuration currently builds all library dependencies rather than
using the "system" ones. We might be able to reuse some of the existing
libraries which already have been ported (e.g. freetype) and split out
those which could be reused by other applications (e.g. icu), but we
would need to solve the dependency management first, in particular the
versioning as TeX Live requires specific versions in many cases.

The script generates a terminal application for each binary which is
only useful for debugging at the moment. We need to provide a driver
(e.g. porting Rubber or latexmk) to setup the pipeline for compiling
TeX documents which could be then integrated with a text editor for
authoring these. However, that should be a separate, standalone
application developed independently of this port.

BUG=
R=bradnelson@google.com, sbc@chromium.org

Review URL: https://codereview.chromium.org/24816002

git-svn-id: https://naclports.googlecode.com/svn/trunk/src@990 7dad1e8b-422e-d2af-fbf5-8013b78bd812