Add leveldb port
This port is not yet fully working or tested.  It is
a good port to add, as it is used by a lot of other
third_party code, including chromium, and bitcoin.

The next step is to get all the tests passing but this
requires mkdir and friends to be wired up the IRT
interfaces in when building for sel_ldr.

TEST=./make_all.sh leveldb
R=binji@chromium.org, bradnelson@google.com

Review URL: https://codereview.chromium.org/136103002

git-svn-id: https://naclports.googlecode.com/svn/trunk/src@1040 7dad1e8b-422e-d2af-fbf5-8013b78bd812