Initial port of Toybox and toysh.
Toybox is a small, busybox-like collection of system tools, including a
simple shell.  The shell here has been modified to avoid
fork/spawn/waitpid for internal commands, so it should be functional for
examining a system at its basic level.  This should work, preliminarily,
with the current version of nacl-spawn in the repository.  Note that a
spawn can be forced by calling "toybox ls" for instance.

This also sets up a top-level PPAPI function within the toysh
executable; because of the odd nature of the build process for toybox,
it may be possible to abstract this out into a standard _ppapi.c
but because it looks like the state of *_ppapi.c functions is somewhat
in flux, I have left that in pending future revisions.

Finally, getcwd seems to only work for PNaCl builds, not any of the NaCl
builds.

R=sbc@chromium.org
BUG=

Review URL: https://codereview.chromium.org/136883007

git-svn-id: https://naclports.googlecode.com/svn/trunk/src@1048 7dad1e8b-422e-d2af-fbf5-8013b78bd812