Add top level build script for building packages.
This takes a lot of the builderplate out of the individual
build.sh script and puts it in a central script:
naclports.sh. This script drives building of packages and
include common.sh and the package specific build.sh.

This is one step towards a more centrally controled build
environment where each step of the build can be run, and
logged separately, making the build output easier on the
eye (and more like system's like Mac's homebrew).

Make the build steps for each package more independent. The
idea being that they could be run as separate processes
in the future.

Also update the top level README to use 'package' rather
than 'port' where is makes sense.

R=binji@chromium.org, bradnelson@google.com

Review URL: https://codereview.chromium.org/132343007

git-svn-id: https://naclports.googlecode.com/svn/trunk/src@1052 7dad1e8b-422e-d2af-fbf5-8013b78bd812