Move dependency information into pkg_info files.
This is a step towards removing the need for a top level
Makefile to drive building of ports.  The dependency
information for each port is now stored in the pkg_info
file.  The top level Makefile no longer contains the
information but generates it automatically using gen_deps.py.

R=binji@chromium.org, bradnelson@google.com, matthewturk@gmail.com

Review URL: https://codereview.chromium.org/143973011

git-svn-id: https://naclports.googlecode.com/svn/trunk/src@1077 7dad1e8b-422e-d2af-fbf5-8013b78bd812