Run executables in HTML5 filesystem by nacl-spawn
Now nacl-spawn can find shared objects which are necessary to
to run the main binary when the main binary is stored in the
HTML5 filesystem. As there are no obvious default library
paths in our HTML5 filesystem, it only finds libraries in
LD_LIBRARY_PATH environment variable.

Then, nacl-spawn constructs a NMF with the result of the
dependency finder. It supports both statically linked binaries
and dynamically linked binaries.

As nacl-spawn does not know the path where the main HTML and
the JavaScript are stored, naclterm.js converts filepaths in
the NMF to fully qualified URLs.

It seems newlib lacks elf.h, so currently we do not support
spawning a new glibc process from a newlib process.

This change also introduces Makefile for nacl-spawn.

TEST=./make_all.sh nacl-spawn
TEST=Manually tested with my local environment whose set up
     script is not landed yet. The commands were:
   $ cd /mnt/html5
   $ export LD_LIBRARY_PATH=mingn/lib:mingn/usr/lib
   $ ./hello64
   $ ./hello64-newlib
   $ ./hello64-static  # glibc static binary, failed.

R=sbc@chromium.org

Review URL: https://codereview.chromium.org/145563003

git-svn-id: https://naclports.googlecode.com/svn/trunk/src@1078 7dad1e8b-422e-d2af-fbf5-8013b78bd812