Link some CLI apps against libcli_main.a
We can reduce duplicated code by this change. The main
motivation of this change was to enable chdir($PWD) from
these apps so that they can be used from mingn. I also
considered modifying ppapi_simple, but to do this, we need
to move mkdir and mount to ppapi_simple as well.

TEST=./make_all.sh \
  python_ppapi python3_ppapi nano ruby_ppapi vim bash
TEST=Some manual tests for the above 6 apps.
R=bradnelson@google.com, sbc@chromium.org

Review URL: https://codereview.chromium.org/134183006

git-svn-id: https://naclports.googlecode.com/svn/trunk/src@1084 7dad1e8b-422e-d2af-fbf5-8013b78bd812