Fix jpeg8b glibc build.
Under glibc, the .nexe file for jpeg8b are created in the
.libs folder and the default post build step creates wrappers
in the same folder.  However, the test code needs wrappers in
the build directory, so we need to create these manually.

Ths change essentially reverts the change made to jpeg8b
in r1110.

R=binji@chromium.org

Review URL: https://codereview.chromium.org/171603003

git-svn-id: https://naclports.googlecode.com/svn/trunk/src@1112 7dad1e8b-422e-d2af-fbf5-8013b78bd812