Fix race condition in chrome_test, re-enable chrome_test tests on the bots.
It turns out that if a message port is disconnected, outstanding messages are
not guaranteed to be delivered before onDisconnect. Changing the behavior of
the reset message to allow better synchronization.

Setting environment variable CHROME_DEVEL_SANDBOX on the bots to allow
plumbing_test to run.

Correcting filename in download code to ensure mac binaries are marked
executable.

BUG=None
TEST=local + on bot
R=binji@chromium.org

Review URL: https://codereview.chromium.org/204793002

git-svn-id: https://naclports.googlecode.com/svn/trunk/src@1169 7dad1e8b-422e-d2af-fbf5-8013b78bd812