Add gtest style filters to chrome_test, multiple test dirs, and refactor TEST() macro.
Adding a -f/--filter option to chrome_test that functions approximately like
googletest filters. This allows you to run a subset of the tests.

Refactoring TEST() macro to be implemented on top of the TEST_F() macro.

Adding support for multiple test dirs (to supported mixing checked in and generated files in a test).

BUG=None
TEST=plumbing_test
R=binji@chromium.org

Review URL: https://codereview.chromium.org/199793013

git-svn-id: https://naclports.googlecode.com/svn/trunk/src@1173 7dad1e8b-422e-d2af-fbf5-8013b78bd812