buildbot: Add missing bzip2 to partition 0 of 3
This dependency was added to freetype by a previous commit

Also, fix build_shard.sh so that is can handle failures
in partition.py in a sane way.

Also update BUILD_ROOT in naclports.py (it was still pointing
to the old 'repository' location).

R=binji@chromium.org

Review URL: https://codereview.chromium.org/211383005

git-svn-id: https://naclports.googlecode.com/svn/trunk/src@1177 7dad1e8b-422e-d2af-fbf5-8013b78bd812