Pass test extension via user agent and make messaging more interlocked.
When running in the test harness, extensions meant to work together may not
know each others ids. Pass the testing extensions id as part of the user
agent so that extensions can use it in testing mode to locate their peers.

Provide reply messages to the 'proxy' and 'setDebugPort' messages so that
they are processed in order. (Seems they are not always).

Changed 'proxy' message to take an extension name instead of an extension id, as these are not fixed.

R=binji@chromium.org
BUG=None
TEST=plubming_test, gdb_test

Review URL: https://codereview.chromium.org/212913004

git-svn-id: https://naclports.googlecode.com/svn/trunk/src@1182 7dad1e8b-422e-d2af-fbf5-8013b78bd812