Split package NAME and VERSION into two fields in pkg_info.
This change adds the version number of the binary package
file. It also adds a consistent naming convention for
binary packages which is:
<name>_<version>_<arch>_[libc]_[debug]

Since we use _ to seperate these components, _ is now
illegal in the package name and package version, just like
debian or ubuntu package names.

R=binji@chromium.org

Review URL: https://codereview.chromium.org/213583002

git-svn-id: https://naclports.googlecode.com/svn/trunk/src@1193 7dad1e8b-422e-d2af-fbf5-8013b78bd812