Build system for statically-linked Python.
This adds a new subdirectory, python_modules/ , and a mechanism for
creating a statically linked Python build.  It is designed primarily for
PNaCl, where a lack of shared library support and the cost of
translation mean that a large Python stack requires a complex,
multi-step build process.  This process can be summarized as such:

 * Build a host python containing all necessary build-time dependencies
   for the stack
 * Build each individual module
 * Collect object files and construct unified libpython2.7.a and
   python.pexe
 * Add a dynload module for statically linked modules

When not automated, this is prone to error and difficult to add or
remove additional modules.  The system included here, while not perfect,
addresses this by enabling individual modules to be selected and built
ala carte, with a reubilding of python_static necessary to create the
unified python.pexe and libpython2.7.a.  The libpython2.7.a output by
this can then be used as a link target for a python_ppapi-like module.

As examples, and as components of a scientific python stack, this
changelist includes:

 * numpy (host and destination builds)
 * cython (host-only build)
 * setuptools (host-only build)
 * matplotlib (dest-only build)
 * sympy (dest-only build)
 * ipython (dest-only build)
 * pyzmq (dest-only build)
 * h5py (dest-only build; requires hdf5, under separate CL review)
 * jinja2 (dest-only build)
 * pyppapi (dest-only build)
 * six (dest-only build)
 * distribute (dest-only build)
 * nose (dest-only build)
 * pyparsing (dest-only build)
 * dateutil (dest-only build)
 * markupsafe (dest-only build)
 * yt (dest-only build)

This CL includes a port for the pyppapi module, which was developed in
support of this and which provides a Python API to communicate between
running Native Client modules and Chrome.  This includes serialization
of NumPy arrays into typed arrays as well as other Python/JS
translations.

Python module ports act very similarly to standard naclports modules,
with two exceptions.  The installation of the .py files has been moved
into the building phase and the collection of .o files occurs during
installation.  Additionally, they utilize modules.libs and modules.list
files.  The first of these files should append to the variable
PY_MOD_LINK to add on additional libraries for building the module, and
the second is a list of compiled modules (not object files) produced by
the python module which should be included in the symbol table in
libpython2.7.a.  For instance, if the compiled C module "fastcode"
exists in a Python module and it requires linking against cfast, these
files would look like:

modules.list:

fastcode

 modules.libs:

PY_MOD_LINK+=" -lcfast"

At present, this has been tested on PNaCl.  To construct a Python build
with a number of modules, a command like this can be issues:

NACL_ARCH=pnacl make numpy sympy matplotlib python_static

The resultant installed libpython2.7.a will contain all of those
modules.  Note that any pure-python modules auto-installed by setuptools
will be functional in the resultant python destination, but any
C-compiled python modules that have been auto-installed will simply
fail.

R=sbc@chromium.org, hamaji@chromium.org, kestert@google.com
BUG=

Review URL: https://codereview.chromium.org/138913004

git-svn-id: https://naclports.googlecode.com/svn/trunk/src@1230 7dad1e8b-422e-d2af-fbf5-8013b78bd812