Switch pnacl to use usr/local for install ports.
Previously we were sharing 'usr' with parts of the toolchain
itself, which meant we couldn't clean the entire folder as
a fast way to remove all ports.

Also, crucially,  it meant that we couldn't hide system
include files and then reference them with #include_next.

The pnacl compiler already include usr/local/include when
searching for headers and there is a change in the pipe
to add usr/local/lib to the library path.  Until that lands
we patch the driver code in-place.

R=binji@chromium.org

Review URL: https://codereview.chromium.org/294363002

git-svn-id: https://naclports.googlecode.com/svn/trunk/src@1288 7dad1e8b-422e-d2af-fbf5-8013b78bd812