Pass --no-deps when building naclports SDK bundle.
This prevents extra libraries (ones that we don't
currently ship) from being built when the bundle is
built.

This fixes the bzip2 optional dependency in freetype.  For
the ports bundle we don't ship bzip2 so we don't want
freetype to built with bzip2 supports.

R=binji@chromium.org

Review URL: https://codereview.chromium.org/303273003

git-svn-id: https://naclports.googlecode.com/svn/trunk/src@1292 7dad1e8b-422e-d2af-fbf5-8013b78bd812