Add initial port of SDL2
This port was contributed largely by jirkadanek7.
For example: jirkadanek7

Although it builds there has been no testing done
on this port so its not expected to work.  This is
simply an initial commit from which we can start
work.  Since the patch is large and we will likely
be doing a lot of work on it we use github to maintain
the patch history.

R=binji@chromium.org

Review URL: https://codereview.chromium.org/242803002

git-svn-id: https://naclports.googlecode.com/svn/trunk/src@1299 7dad1e8b-422e-d2af-fbf5-8013b78bd812