Add the ability to view graphical output of programs within the dev environment.
This change adds the ability to view the graphical output of programs
(from Pepper APIs) within the dev environment in either a popup window
or as a section of the terminal window. This is enabled by setting
the PROCESS_SPAWN_MODE environment variable to either 'popup' or 'embed'.
The dimensions of the viewing area can also be configured.

BUG=none
TEST=none
R=binji@chromium.org, bradnelson@google.com

Review URL: https://codereview.chromium.org/349453004

git-svn-id: https://naclports.googlecode.com/svn/trunk/src@1397 7dad1e8b-422e-d2af-fbf5-8013b78bd812