Add requests and fix dependencies for thingking.
Thingking requires 'requests', which if not installed, will be
downloaded by setuptools.  This prevents that from occuring, and
should fix the build on OSX, where SSL downloads are not possible.

Additionally, this removes functools32 from the requirements
for thingking as that requirement is fulfilled by yt instead.

BUG=
R=binji@chromium.org, kestert@google.com

Review URL: https://codereview.chromium.org/477473003

git-svn-id: https://naclports.googlecode.com/svn/trunk/src@1536 7dad1e8b-422e-d2af-fbf5-8013b78bd812