Update SDL version to 1.2.15.
Several chucks of the nacl.patch file are no
longer needed as they are upstream in this version.

This is in preparation for switching to github
for SDL patch maintenance.

Also, update scummvm patch so that it doesn't
try to mkdir(/home) which is already mounted
by SDLmain.

TEST=build and run scummvm and snes9x
R=bradnelson@google.com

Review URL: https://codereview.chromium.org/502063003

git-svn-id: https://naclports.googlecode.com/svn/trunk/src@1572 7dad1e8b-422e-d2af-fbf5-8013b78bd812