Fix build of 'six' python module.
When this module was built in isolation it was
failing because the python site-packages directory
was not yet created.

This change creates this directory in the
DefaultPythonModuleBuildStep.

As a most long term solution we should have all
python modules depend on the python package but
we currently have conflict with the python-static
package.

TBR=bradnelson

Review URL: https://codereview.chromium.org/549673002

git-svn-id: https://naclports.googlecode.com/svn/trunk/src@1627 7dad1e8b-422e-d2af-fbf5-8013b78bd812