Switch elf_reader to read both 32 and 64 bit elf files.
As a PNaCl build of elf_reader will not know which arch it will run on at
compile time, it needs to be able to cope with both kinds of elf files (32/64).
Switching it to read initially into EHDR32 (which has the same layout
regardless of bitage), then switch to use the proper struct size dynamically.
Publish 64-bit structures for maximum generality, copying the fields.

BUG=None
TEST=devenv_tests
R=binji@chromium.org

Review URL: https://codereview.chromium.org/551973003

git-svn-id: https://naclports.googlecode.com/svn/trunk/src@1640 7dad1e8b-422e-d2af-fbf5-8013b78bd812