Fix openssl build on OSX
openssl's config script would use `uname` to try to
figure out the build configuration.  For the NaCl
build we instead want to explicitly specify $MACHINE
$SYSTEM and $RELEASE.

This change allows openssl to build on Mac OSX as
all the packages the depend on openssl.

Also, change the was the broken ssl tests are filtered
out in the newlib build because the sed invocation was
using a gnu-sed-ism which wasn't working on OSX.

R=bradnelson@google.com

Review URL: https://codereview.chromium.org/549393002

git-svn-id: https://naclports.googlecode.com/svn/trunk/src@1645 7dad1e8b-422e-d2af-fbf5-8013b78bd812