Switch from using 'git number' to 'git describe'
git number is non-standard (its part of gclient)
and not fully supported.  The reason we were
using it is because it doesn't rely on the
existence of the git annotated tag.  Now that
naclpors has officially switched to git we
can push annotated tags to the master repo and
start using 'git describe' to label builds.

BUG= https://code.google.com/p/naclports/issues/detail?id=139
R=bradnelson@google.com

Review URL: https://codereview.chromium.org/563603002