Pulling in newer package versions into devenv, holding back some.
Pull in newer versions of everything other than git and python.

Fixing link order in git that causes it to malfunction (gets bad access on working directory). Need to roll forward in a later version.

Python is a mystery.

Advancing the version number of devenv to allow it to be published in the store.

Fixing help in packaging script + allowing pinning.

BUG=None
TEST=tested video demo
TBR=sbc@chromium.org

Review URL: https://codereview.chromium.org/566183003