Fix ncurses build.
Setting ARFLAGS was causing confusing the configure
script for ncurses. It tries to combine ARFLAGS with
a selection of possible other options, no of which
work which ends up loading "unknown" in the final
ARFLAGS in the Makefile.

This change effectively pre-seed the cf_cv_ar_flags
causing ncurses to skip the flag discover phase.

R=binji@chromium.org

Review URL: https://codereview.chromium.org/618813003