Fix jpeg6b package
Previously we were disabling the install of this
package to avoid conflicting with jpeg8d but
now we have a better way of dealing with this
(CONFLICTS in pkg_info).

This change means that opencv will correctly detect
and use the jpeg6b package.

BUG= https://code.google.com/p/naclports/issues/detail?id=107
R=binji@chromium.org

Review URL: https://codereview.chromium.org/612383005