Fix compiler used for boost build.
A recent change caused the user-config.jam to be written
incorrectly and therefore the wrong compiler to be used.

Turns out we need to disable a few more parts of boost
that previously thought in order to get a build (at
last for ARM and PNaCl builds).

We really should be building and verifying some of the
boost examples/tests.

R=binji@chromium.org

Review URL: https://codereview.chromium.org/622513003