Use sha1 rather than timestamps when checking build directories are up-to-date.
Previously naclports would determine that a working build
directory was out-of-date if the patch file was more recent
that the stamp. Instead store the SHA1 of the patch in the
stamp and check the stamp contents.

This is more robust and leads to fewer false-positives (for
example if when switching between branches that have
different patch contents).

R=binji@chromium.org

Review URL: https://codereview.chromium.org/626543002