Fix python3 glibc build
The build was failing because the PYTHONPATH used
to run python commands during the build was including
the poth to the shared objects for the cross build
resulting in failure to load the .so files.

IIUC we can't actually ever load any .so files from
the cross build, so I removed this path from the
PYTHONPATH used during the build.

TEST=TOOLCHAIN=glibc make python3
R=bradnelson@google.com

Review URL: https://codereview.chromium.org/650883003