Replacing xfree86 keymap with one that matches SDL + PPAPI.
This makes the keyboard usable in xorg-server and in emacs!

So this file was intended to map from 102 key keyboard up/down codes to a vitual notion of keys in X11, mind you this is before deciding the particular layout, so its more like the physical position in row and column.

SDL forwards what it thinks is the raw keycode, in nacl this comes from PPAPI.

PPAPI in turn defers the list to ... Windows:
 /**
   * This value reflects the DOM KeyboardEvent <code>keyCode</code> field.
   * Chrome populates this with the Windows-style Virtual Key code of the key.
   */

So I've replaced the default keymap based loosely on this (but in decimal):

http://msdn.microsoft.com/en-us/library/windows/desktop/dd375731(v=vs.85).aspx

I've commented out the ones that I don't have an equivalent for, because my observation is that colliding key values cause the last one to win (and some of the later entries have the same code as earlier ones).

Also I've deleted a large chunk at the end for Japanese keyboards, as these also collide, and this format does not seem to support /* */ style comments.

BUG=None
TEST=local
R=petewil@chromium.org,binji@chromium.org

Review URL: https://codereview.chromium.org/679883002