Add boringssl port
Currently only builds with glibc (due to horrible
issue in newlib headers when building with -std=c89).

Tests build and run as part of the build.

Some parts of this patch can most likely be upstreamed
to boringssl.

Landing with DISABLED=1 since the version of cmake required
is not available on the buildbots (Ubuntu/Precise).

BUG=
R=bradnelson@google.com

Review URL: https://codereview.chromium.org/682633004