disable use of getrlimit (and other fixes for new libnacl)
This set of changes is in preparation for the
recent libnacl changes which added a lot of
new stub functions.

Because of the new stubs configure will detect
more thigns which can sometimes this can lead
to sources that don't compile.  For example the
getrlimit/setrlimit function are now available at
link time, but corresponding macros for calling
them are not yet in the headers.

Remove SIN6_LEN from glibc-compat in6.h.  This macro
should not be there unless sin6_len field is part
of the struct.

R=binji@chromium.org

Review URL: https://codereview.chromium.org/702153003