Adding jseval utility for use in open web devenv.
Eval is allowed on the open web, and when used inside the devenv
can compensatate for some of the limitations of the web pages
(sockets + wjere to emit things to).

Adding a pnacl only utility that allows a page to request javascript
to be run synchronously.

Adding an environment variable containing the current origin.

BUG=None
TEST=local,trybots
R=binji@chromium.org

Review URL: https://codereview.chromium.org/719553002