Don't export NACL* tool variable in nacl-env.sh
nacl-env.sh is designed to be sources from other
scripts and should not export anything iself.
Instead export these variables in common.sh.

Also, move the code for disabling the .la files
that ship in the SDK from nacl-env.sh to
common.sh.

This change is in preperation for converting
nacl-env to python.

R=binji@chromium.org

Review URL: https://codereview.chromium.org/703293003