Allow ctrl-c to pass thru naclterm if sigint doesn't apply, enable emacs.
Allow ctrl-c to pass thru to application when ENABLE_CONTROL_C is not
turned on, or when there's only one process.

Disable ENABLE_CONTROL_C for emacs in the devenv.

Add a special case for the emacs package to match the existing layout.

Switch emacs to common untar function to allow it to run inside
the devenv.

BUG=None
TEST=manual + bots
R=binji@chromium.org, petewil@chromium.org

Review URL: https://codereview.chromium.org/763353005