Don't rely on fine-grained file timestamps
The patch step was being applied if its timestamp
was not newer than the extract timestemp.  On OSX
the two timestamps were aften equal since the two
events would occur within the same second.

Revert to the previous behavior which was to remove
the patch stamp each time a new archive is extracted
or cloned.

R=binji@chromium.org, bradnelson@google.com

Review URL: https://codereview.chromium.org/790603002