ninja: update to 1.5.3 and implement out-of-tree build
This allows the host ninja to be built just once
and then all the target ninjas to be build in
their own tree.

Add ninja testing (with subprocess tests disabled)

Fix the type of rlim_t in glibc-compat to be unsigned
(in accordance with POSIX) to fix a warning in the
ninja unit tests.

R=bradnelson@google.com

Review URL: https://codereview.chromium.org/797793002