Add 'naclports uscan' command which checks for new upstream versions.
This new command piggybacks on Debian's 'uscan' command
to determine if there are new upstream releases available
alongside the package's URL (This means that this command
won't work on Windows or Mac).

It requires a packages URL to be modified to unclude ${VERSION},
which can then be replaces with '(.+)' so uscan can search
for newer versions.

Tested so far with just imagemagick. A followup to this
change will modify the pkg_info for all packages to use
${VERSION} in the URL field.

R=bradnelson@google.com

Review URL: https://codereview.chromium.org/828903003