Restore quiet logging during package build (by default)
This restores the old behavior of minimal output
except in the case when --verbose is specified.

When --verbose is not specified, and we log to a file,
verbose logging is enabled while writing to the file.

Also, update log file name to include full configuration
name so different builds of the same package write to
different log files.

R=bradnelson@google.com

Review URL: https://codereview.chromium.org/840773003