Update pinned version of chrome used for testing
Also download the PNaCl translator component along with
chrome so that we can test binaries built with that latest
toolchain without having to wait for the public component
to be updated.

This should fix the current PNaCl test failures on the bots
which are due to the latest toolchain generated pexe files
that the publis PNaCl component cannot yet handle.

R=binji@chromium.org

Review URL: https://codereview.chromium.org/829883003