glibc-compat: Add lockf() and some constants required by cups
This CL implements the lockf() method and adds some constants. Both
are required to compile cups. Note that only the lockf() commands
are implemented which can trivially be tested. If the F_TEST command
is needed in the feature, it will need to be added.

R=sbc@chromium.org

Review URL: https://codereview.chromium.org/868703003