Disallow the use of __uint128_t in cairo for nacl-clang
nacl-clang declares __uint128_t (but not uint128_t), which
is a gcc extension. However, it does not provide a runtime
support for neither __uint128_t nor uint128_t in compiler-rt
which causes an error during linking. This is a temporary
workaround until nacl-clang is fixed.

BUG=https://code.google.com/p/nativeclient/issues/detail?id=4112
TEST=NACL_ARCH=x86_64 TOOLCHAIN=clang-newlib make cairo
R=sbc@chromium.org

Review URL: https://codereview.chromium.org/988983002