Remove nacl_io_uninit call from cli_main
This is no longer needed now that ppapi_simple no
no longer initialised nacl_io when running under
sel_ldr.

This code was also causing an assertion failure
in ppapi_simple but this is only a problem for
debug builds so it wasn't seen in the builder.

R=binji@chromium.org

Review URL: https://codereview.chromium.org/991993002