Fix status message when uninstalling packages
When an existing package is removed log the infomation
on the existing package, not the information on the
package to be installed.

Also, don't check NACL_DEBUG directly in source_package.py.
Instead rely on the configuration object to contain
the correctly debug status.

BUG= https://code.google.com/p/naclports/issues/detail?id=198
R=binji@chromium.org

Review URL: https://codereview.chromium.org/1008893003