Build bzip2 command line tool for the devenv.
Previously we only built the bzip2 library for use in other programs.
The bzip2 command line tool may be useful in the nacl devenv,
adding it to the build.

NOTE: The newlib executable partially functions (command line parsing
and help output), but fails when compressing.
The glibc executable works.
TODO added for this.

Adding a devenv package for bzip2, lock it to glibc for now.

BUG=None
TEST=manual
R=sbc@chromium.org, gauravashish07@gmail.com, khim@chromium.org

Review URL: https://codereview.chromium.org/1009583005