Only use cached SDK download if full URL matches
Previously, in download_sdk.py we considered out/nacl_sdk
up-to-date if its version matched the requested version.
This was causing issues for the bionic builders that download
the bionic variant of the SDK, which has the same version
but has a different filename.

Review URL: https://codereview.chromium.org/1043503002