Remove lua from the SDK bundle
It was causing the SDK bundle build to fail since it now
depends on cli_main, which we don't want in the SDK.
What is more the SDK bundle is hopefully being phased out
soon anyway.

This allows us to also remove the 'basic' lua build
which didn't rely on readline.

Review URL: https://codereview.chromium.org/1142583005