Converted naclports CQ config to proto-format
Validation will be done via an endpoint in the luci-config.appspot.com app,
which is being implemented at the moment. For now, I've removed the CQ test
because currently it would require a copy of the compiled proto library, which
will get out of date any time the cq.proto is changed.

This will land in sync with https://chromereviews.googleplex.com/200267013 and
should be reverted in sync if needed.

R=akuegel@chromium.org, bradnelson@chromium.org
NOTRY=true
NOTREECHECKS=TRUE
BUG=471767

Review URL: https://codereview.chromium.org/1154423008