buildbot: Uninstall all packages for all architecures between builds
Previously the builtbots would uninstall all packages
for the each arch as then built a package for a given
architecture.  However, due an issue with clang-newlib
toolchain sometimes packages install for x86_64-clang
and effect packages for i686-clang (See:
https://code.google.com/p/nativeclient/issues/detail?id=4108)

This change means that we uninstall everything for
all architecures before building anything.  The specific
problem this was causing is the sdl2 when built for i686
was compiling against the libiconv headers from x86_64
package.

Review URL: https://codereview.chromium.org/1183923002