Fix glibc build of xorg-server
This was broken by the recent removal of ioctls
from the glibc build (since we now override bits/ioctls.h
for glibc to handle the fact that its empty in the new
arm/glibc toolchain).

This patch is now cleaner and more likely suitable for
unstreaming.

Review URL: https://codereview.chromium.org/1256023005