Fixing devenv mounter formatting adding dots.
The mounter ui shows up broken when viewing a web based version of
the devenv. Additionally, the formatting of the mount ui does not
show up correctly on desktop linux.

Correcting the formatting by switching to fixed instead of absolute layout.
Fail gracefully if local storage is not available.
Adding ... markers to the mounter for better discoverability.

BUG=None
TEST=manual
R=gdeepti@chromium.org

Review URL: https://codereview.chromium.org/1269843002 .