Add "which" utility script to devenv
This script is taken from the ubuntu debainutils package.
In the long run we could package the whole thing.

Also, fix bug handling of shbang script where there
is a space after the bang.

Sadly `which which` doesn't work right now because
the script lives on the http filesystem which doesn't
report files as executable.

R=bradnelson@google.com

Review URL: https://codereview.chromium.org/1258333004 .