nacl_startup_untar: only extract when tar conents change
When the extraction target is a persistent filesystem
it can be useful to only extract the tar archive if
its contents has changed.  The method will skip the
extraction iff a file called <tarfile>.hash exists
and the contents match a file the target root.

If either hash file doesn't exist, or if the contents
don't match we proceed with (re)-extraction.

This change doesn't have much effect right now since
most of these apps run with a memfs (ephemeral) root.
However it will be required for the pending change
to use html5 as the root fs in devenv.

In the long run we will move the contents of the tar
in the devenv package, but initially we want to continue
to support nacl_startup_untar in devenv.

Review URL: https://codereview.chromium.org/1278503003