Add font support for GTK+
Updated dependencies for pango, and fontconfig (as well as added new pango
dependency harfbuzz). This was in order to add fonts to the gtk+ demo. Using the
nacl_startup_untar method to insert the fonts, font.conf, cache and various
alias's and modules. Currently it only runs for the glibc toolchain.

BUG=None

Review URL: https://codereview.chromium.org/1262413002