Build pkg files from result of InstallStep
Previously we were only building pkg files for
packages the called PublishByArchForDevEnv.  The goal
is to remove PublishByArchForDevEnv and just rely on
the InstallStep of the package, but for now we build
two different packages in this case: <name>.tbz and
<name_publish>.tbz.

Tested with bzip2 and vim packages.

Review URL: https://codereview.chromium.org/1282653003