Add corelibs package
The library includes the core NaCl SDK libraries and
headers.  It is designed to be installed in dev env, and
not into the toolchain (which already contains these
libraries).

This package is needed as part of the transition from
ad-hoc packaging to freebsd pkg package.  All packages
will essentially depend on this one. The mingn package
will then builds on this to include developer tools.

TEST=./make_all.sh corelibs
TBR=bradnelson (since we are in hurry to switch to pkg before zhiting finishes)

Review URL: https://codereview.chromium.org/1294283003 .