Make Emacs build with clang-newlib
The "pselect" function is not supported in newlib.  Glibc supports
pselect by calling select under the covers, so this change will
explicitly call select when building emacs.

The change as written will call it unconditionally, it might be
better to only do this if we are building newlib, but I wasn't
sure which #define symbol to use for newlib builds.  I'll happily change
the code if you can tell me a better symbol than NACL_EMACS.

BUG=
R=bradnelson@chromium.org

Review URL: https://codereview.chromium.org/1317263005.