Fix host build of pkg in the absence of liblzma
The host build pkg is used to build the repo metadata
files. When building on the bots liblzma is missing
and there is a bug in pkg that ignores the
packing_format field in the meta file.

This change passed --without-lzma when building
libarchive and patches pkg to handle alternative
packing_format in the meta file.

TBR=bradnelson

Review URL: https://codereview.chromium.org/1326273002