Improve emscripten download, install and auto-detection
This change moves the emscripten installation testing
and building of the native optimiser out of the buildbot
script and into the download_emscripten.py script.

It also moves the setting EMSCRIPT and EM_CONFIG out
of the builtbot script and into util.py where it can
be used by the normal development workflow.

R=binji@chromium.org

Review URL: https://codereview.chromium.org/1393343002 .