Fix link error with busybox and glibc
A recent change to ppapi_simple in the SDK caused the link
to start failing.  This change fixes the issue and also
updates the build system so that all settings are passed
in at defconfig time, and also switches to using nacl-spawn
rather than ppapi_simple directly.

Also, an a simple test.

Review URL: https://codereview.chromium.org/1418623008