Build gtest from source rather than using SDK-provided version
The SDK-provided pre-built version of gtest is being
removed and replaced with a source-only version which
is more in line with the recommended practice.

Update gtest revision to the same one currently used
by chrome in order to minimize any code changes.

gtest 1.8.0 is still not out but we can/should update
to that once it is.

R=binji@chromium.org

Review URL: https://codereview.chromium.org/1555913002 .