Update patch set 2
Patch Set 2:

i'm pretty sure this failure is because the python cross-compile is broken.  upstream has long gotten it wrong (or not even tried), so unless nacl is trying beyond that, it's just broken.  if python is mixing build & host files, it might because the build has an older zlib than the nacl, and the headers are diff.

i say this because i've done a lot of python cross-compile work in Gentoo & CrOS so i know upstream is bad :).

Patch-set: 2
Reviewer: Mike Frysinger <1000768@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>